Life expectancy, mean years schooling, and expected years schooling implemented.
- Add a new class shared between server and client,
GHumanDevelopmenUtilities, that packages up the extra data into the
GEventCountryInfo that's passed from server to client.
- Grammatical fix to the covert actions mission outcome window, in the
case that the mission succeeds but the attacker is found out.
- Call GWorldBehavior::LimitExportsToProduction() in a different place
within ExecuteMarket(), which seems to the fix infinite loop issue where
LimitExportsToProduction() keeps getting called without making progress.
- Fix the GWorldBehavior::VerifyEconomicFailure() log to have a $ sign
in front of the debt expenses.