Various changes and refactoring.
- Replace all references to SP2-C with SP2-HDM.
- Display the income index in the economic health window, next to the
GDP per capita value.
- Remove the resources window shortcuts to set all resource tax rates
and to set all exports to their maximum.
- Modify income tax revenue calculation to better reflect real-life
income tax revenue patterns.
- Some refactoring of country-specific HDI calculation, no functional
changes there.
- Modify world average HDI calculation so that it truly treats the whole
world as one country, independent of how many countries actually exist.
- Add a commented-out utility to log out each country's HDI and
components every 90 days of game time.
- Change the AI war logic:
- When the AI creates a war declaration, it will invite all countries
that can join the war on the attacking side, based relations with the
attackers and defenders; before, not all those countries would be
invited, as the game would have a limit on the total military strength
of the attacking forces.
- A country that's considering joining a war on the defending side will
be more willing to join even if the total defending military strength
wouldn't be high. This is no longer dependent on the AI aggressiveness
level.
- A country that's considering joining a war on the attacking side will
be more willing to join, even if the total attacking military strength
wouldn't be high. This is now dependent on the AI aggressiveness level.