V2.
- Internet multiplayer now works. However, some minor UI-related
features had to be removed in order to get it to work.
- Every 365 days of game time, the HDI, LE, MYS, and EYS of every active
country will be written to the server console log.
- Fixed an issue where during the logging out of stats to the server
console at the beginning of the game, countries with 0 population would
have their GDP per capita calculations messed up; such countries would
now be listed as having $0 GDP per capita.
- Check in the ModInfo.xml.

- To implement working Internet multiplayer, the Server.dll had to be
made to work with an unmodified Client.dll. Features that rely partially
or totally on client-side changes had to be removed, which were limited
to:
- In the country information window, your numerical HDI, LE, MYS, and
EYS will no longer be shown.
- In the economic health window, your "income index" will no longer be
shown.
- If your covert mission succeeds but you're found out as the
perpetrator, then you will no longer be informed as such in the mission
outcome window.