V5.
- Military units that are hostile against each other will start battling
each other from a farther distance; this negates the tactic of
excessively splitting one's military units to make it artificially
difficult for hostile units to engage in battle. Thanks to Fenix for the
code suggestion.
- Added a configuration setting to limit the number of covert cells that
a human player may assign to any one foreign country. Thanks to Fenix
for the suggestion. The limit does not apply to AI-controlled countries.
- Fixed an issue in which if a country is conquered, and then another
country changes its name, then the conquered country would reappear in
country listings.
- When performing covert actions against a target country, the positive
effect on success rate of having additional cells in the target country
has been lessened. Thanks to Fenix for the suggestion.
- Added a configuration setting to force the AI to always refuse Assume
Debt treaties when an AI country is requested to assume a human player's
debt.
- Added a configuration setting to control the effective damage radius
of nuclear missiles.
- Added a configuration setting to control how much of a region's
resource production is lost upon being annexed. The production loss does
not affect regions gained through trade.
- Added a configuration setting for dedicated servers, to choose whether
to autosave to the same folder as joshua.exe.
- Added configuration settings for adjusting the expected (not actual)
stability levels required for entering and exiting Anarchy.
- Minor modifications to the Server.vcxproj file.
- When a nuclear-armed AI country has enough of its regions under
foreign occupation, it will launch nuclear attacks against the
occupiers' military, instead of against the occupiers' civilians. Thanks
to Fenix for the suggestion.
- Fixed a potential issue in which countries with extremely low revenue
wouldn't be able to build or buy non-infantry units.
- A nuclear-armed AI country, in response to a nuclear launch involving
another country that it really likes or really dislikes, will have a
bigger preference for itself to launch a military nuclear strike rather
than civilian. Thanks to Fenix for the code suggestion.
- When starting a new game, each country's population will be logged out
to the console in addition to all the stats already being logged.
- When a country conquers another country in which the conquering
country has covert cells stationed, then the conquering country will
have its national security correctly recalculated immediately.
- When a country gains a region, it will lose stability based on the
ratio of the region's population to the gaining country's population.
The stability penalty is lower if the region is gained through trade.
- Nuclear strikes against civilians now affect regional populations that
are not attached to any specific cities. This effectively increases the
amount of people that nuclear strikes can kill at once. Thanks to Fenix
for the suggestion.
- Nuclear strikes against civilians now affect telecom development level
as well as that of infrastructure.
- The Free Region treaty now only gives relations bonuses between the
original owner of the region(s) and the former occupier of the
region(s), instead of between everyone and the former occupier. Thanks
to Fenix for the suggestion.
- Changed how the game determines whether a covert mission succeeds,
whether or not the true perpetrator would be found out, and whether or
not the executing cell would be captured in case of mission failure.
Previously, the outcome of a mission would always be the same as long as
all the mission details (e.g., type, cell training level, complexity,
etc.) and the in-game date and time of mission execution were the same.
- When an Assassination mission succeeds against a target country, if
the target country has any covert cells stationed in itself, then one of
the foreign cell may be removed (assassinated) at random, depending on
the executing cell's experience level and the mission's complexity.
Thanks to Mr Monday for the suggestion.
- It is no longer possible to perform Espionage missions against one's
own country for economic benefit. Thanks to Fenix for the code
suggestion.
- Changed the method of creating 10 covert cells at once. In the cell
name field, type "MULTIPLE, followed by a space, followed by a whole
number; this will create 10 cells, named in sequence according to the
whole number that you specified. For example, if you type "MULTIPLE 21"
into the name field and confirm, you will create 10 cells, named "21",
"22", "23", ..., "30".
- Changed the mod's server list display name to "Human Development Mod
V5" from "SuperPower 2 Human Development Mod".
- When the game saves, either via the dedicated server autosave feature
or the "save" console command, the game will now be saved correctly even
if the "save" folder doesn't already exist. Thanks to Fenix for the
suggestion.
- Fixed an issue where the "set_admin_country" console command wasn't
recognized.
- Added a "print_players" console command to list all active players.
Thanks to Fenix for the suggestion.
- It is no longer possible to move ground units to Antarctica when the
Naval Rule is enabled; this prevents circumventing the Naval Rule by
moving non-naval units to Antarctica before moving them to their final
destination.
- Fixed at least some instances of the issue in which units that are
deployed close to their capital may get stuck in deployment status.
- Unemployment now changes more gradually than before.