Implement a configuration option to prevent defending countries from attacking the home territory of countries on the attacking side. Variou
- Added a configuration option to disallow a country on the defending
side of a war from occupying the home territory of a country on the
attacking side, unless the occupying country is also currently in an
offensive war against the owning country. Region liberation and naval
bombardment are still allowed.
- Modified the GDZDebug log formatting in case no function name and/or
line number were provided.
- In GDataControlLayer::ExecuteMission(), don't create a whole copy of a
country's list of covert cells when examining it.
- Fixed a compilation error in logging instances of countries going
bankrupt.