Further fix to the issue where cancelling in-production infantry can artificially increase a country's population.
- When a country has only 8-9 manpower available for infantry, the
country may still enroll 8 infantry at a time without having that amount
removed from its population. Then, cancelling the enrolment can increase
the country's population artificially.
- Changed GDataControlLayer::RemovePopulationFromCountry() to return the
amount of population able to be removed. If no population was able to be
removed due to enrolment, then enrolment may not proceed.
- The reason the issue happens is because the game tries to remove
population equally from a country's regions; when trying to remove just
8-9 people, no population ends up being removed due to not being able to
remove fractional people, but the enrolment is made to succeed anyway.
In the future, if we can't remove enough for enrolment, we could try to
remove 8-9 population from the first region that has enough population.
- Disabled all DZDebug logs by default.
- Removed an unused variable warning in GUnitMover::MoveUnits().