Unit movement - Issue #46.
- Fixed an issue where deploying units to a capital, while having other
units ready to be deployed, would cause the in-deployment units to
disappear after deployment. This issue had existed in SP2 V1.5.1.
- This was due to the game automatically merging units when a unit is
deployed in the same location as an existing one, but the game sometimes
mistakenly tries to merge a deployed unit with ready-for-deployment
units, which doesn't make sense.
- Fixed another issue where the game's unit merge feature attempted to
create a 2-point unit movement path with the points being the same, but
calculated an indeterminate distance between them, so the units would
disappear from the globe due to being in an invalid state.
- Added GDZLOG()s and gassert()s in the unit movement math to catch
these instances in the future.
- Changed the mod name to V8 from V9. No functional changes.
- Disabled code optimization and enabled frame pointers for
Debug_In_Release, to facilitate better debugging.