Issue #29, AI covert actions against wrong country.
- The AI code assigned cells to the target without checking if the
target was conquered.
- When a country is conquered, redirect all cells that are headed to
that country, to the conqueror. Cancel all unexecuted missions agains
the conquered country and assign them to the conqueror.
- Automatically cancel any mission, by human or by AI, that targets a
conquered country, upon mission execution. Send that cell back to its
home country. This might be triggered if a saved game with an unfixed
#29 is loaded in such a state, or if an AI determines a target before
it's conquered but doesn't start moving cells until after it's
conquered. The latter will be fixed in #23.
- Added GDataAccessLayerServer::CovertCellInfoForLog() to log out cell
info for debugging.