Unit Groups - Followup to #53 and #54 fix.
- Followup to commit 06f8606976ab71bb25f8e818ff74d279c8ac093a, Mar 17,
2017. Don't force NextStatus change for a unit group that just left a
battle that had interrupted its state change, if the UnitMover no longer
registers it as changing state.
- If a group successfully joins an interrupting battle, then its
movement would be cancelled, though its saved NextStatus would still
signify what it would be if it successfully changed status.
- The original GUnitGroupEx::LeaveCombat logic would set NextStatus to
the saved NextStatus as long as it were different from what the
superclass LeaveCombat would set, which would hit a gassert when the
UnitMover finds that the group's current and next statuses are different
despite not changing status.
- Create GUnitMover::IsUnitGroupChangingStatus for
GUnitGroupEx::LeaveCombat to check if it's really changing status. If
not, then LeaveCombat would set its NextStatus to be the same as its
current status.