Economy - Correct and simplify GDP calculation.
- GDP is now equal to sum of all resource production in a country,
without any other modifiers. Doesn't make sense to have GDP scaled due
to gov't type, and based on the way the game calculates GDP as being the
sum of all resource production, it doesn't make sense for imports and
exports to factor in.
- Effectively, GDPValueBase and GDP are now the same, and
EconomicActivity is unused, at least outside the libraries.